requirements.txt
requests
functions-framework
python-dotenv
beautifulsoup4
selenium
webdriver-manager
